VacEngine.EctoHelpers (Vac Engine v1.0.3)

Set of utilities to help manipulate data in Ecto Schema.

Functions

accept_array_or_map_for_embed(attrs, key)

Convert %{foo: %{obja}, bar: %{objb}} into [%{name: :foo}, %{name: :bar}]

delete_all(q)

Delete all with result wrapping

get_in_attrs(attrs, path, default \\ nil)

Used to get data in attributesm will try atom and string keys.

put_in_attrs(attrs, path, value)

Used to put data in attributesm will try atom and string keys to be compatible with existing data.

set_positions(attrs, key)

Set position recursively

shift_position(changeset, group_field, group_value)

Helper to manage position shifting

transaction(multi, key)

Wrapper around Repo.transaction that will return an ok or error tuple

wrap_in_map(attrs, key, target_name)

Wrap an attribute key into a map with the given root key