VacEngine.EnumHelpers (Vac Engine v1.0.3)

Set of Enum helpers

Functions

compact(map)

Remove nil AND empty lists from map

find_by(list, key, value)

Find object in list by given key (list items must be map)

flatten_map(m, stack \\ [])

Convert a nested map to an array of tuple with dotted path

get?(map, key)

Nil safe get, return nil if map is nil

get_mixed?(map, key, default \\ nil)

Get mixed first atom then string, nil safe

sdmerge(a, b)

Simple deep merge of map of map

unflatten_map(a)

Convert a flat array to a nested map