VacEngine.PipeHelpers (Vac Engine v1.0.3)

Helper for piping data

Functions

func_inspect(value, fnc)

Inspect the element returned by a callback function then pass the input value.

func_inspect(value, fnc, label)

Same as above, with a label on the previous line.

noreply(val)

ok(val)

Wrap into ok tuple

pair(val, res)

Wrap into tuple pair

puts_then(value, message)

Put a message on the terminal then pass the input value.

rpair(val, res)

Wrap into tuple rpair

tap_ok(result, fun)

Tap only if ok tuple

tap_on(result, value, fun)

Tap only if value match

then_ok(result, fun)

Then only if ok tuple

then_on(value, value, fun)

Then only if value match