VacEngine.Processor.Library.Functions (Vac Engine v1.0.3)

Actual blueprint expression functions

Functions defined in this module correspond to the functions available to compose complex blueprint expressions.

Function have:

a name (the actual function)

a short name, defined by a @short attribute

a label, defined by a @label attribute

one or more signatures, defined by a @signature attribute. Each signature defines a valid type combination of argument and output

one or more implementations to handle all the possible types and inputs

Functions

add(a, b)

Add two values

add_days(date, days)

Add days to date

add_hours(date, hours)

Add hours to date

add_minutes(date, minutes)

Add minutes to date

add_months(date, months)

Add months to date

add_seconds(date, seconds)

Add seconds to date

add_weeks(date, weeks)

Add weeks to date

add_years(date, years)

Add years to date

age(_)

Calculate age in years given a birthdate

Value returned in years.

and_(a, b)

Boolean AND

days_between(a, b)

Return the day difference between two dates

Always return a positive number.

div(a, b)

Divide two values

earliest(a, b)

Return earliest date

earliest(a, b, b)

Return latest date

eq(a, b)

Check equality of two expressions.

gt(a, b)

Check if a is greater than b. For date, a is after b.

gte(a, b)

Check if a is greater than or equal to b. For date, a is after b.

hours_between(a, b)

Return the hour difference between two dates

Always return a positive number.

is_date(a)
(macro)

Check if struct is date

is_false(arg1)

Check if boolean is false

Returns true if the boolean is nil

is_nil_(arg1)

Check if variable is nil

Returns false only for nil

is_not_nil(arg1)

Check if variable is not nil

Returns true only for nil

is_true(arg1)

Check if boolean is true

latest(a, b)

Return latest date

latest(a, b, b)

Return latest date

lt(a, b)

Check if a is less than b

lte(a, b)

Check if a is less than or equal to b

minutes_between(a, b)

Return the minute difference between two dates

Always return a positive number.

months_between(a, b)

Return the month difference between two dates

Always return a positive number.

mult(a, b)

Multiply two values

neq(a, b)

Check inequality of two expressions.

not_(arg1)

Inverse boolean

now()

Get the current time

or_(a, b)

Boolean OR

seconds_between(a, b)

Return the second difference between two dates

Always return a positive number.

sub(a, b)

Subtract two values

var(name)

Get a variable value

weeks_between(a, b)

Return the week difference between two dates

Always return a positive number.

years_between(a, b)

Return the year difference between two dates

Always return a positive number.