VacEngine.Simulation (Vac Engine v1.1.1)

Provides the model for the blueprint simulation. Simulation is used to assess whether a ruleset meets business requirements.

A blueprint simulation is a collection of input/output definitions. The output corresponds to the expected outcome of the blueprint for the given input.

Simulation runner

The input descriptions are processed a case runner (similar to the main processor) in order to determine if the corresponding output expectations are met.

Stacks, layers, cases

The description passed to the runner is actually a "stack" of simple descriptions called "cases" where the higher case's values overwrite the lower case's ones. Cases are refenced as "layer" in the "stack" and may be referenced in several stacks. Only input is composed this way. Lower cases are not intended to describe output.

Two-layers "Basic stacks" and "Templates"

In the current implementation of the simulation. Stacks are limited to two level. The bottom layer is called a "Template" where the top one is called a "Runnable case".

Blueprint templates and basic stacks

Each blueprint simulation has a collection of templates (a many to many relation between cases and blueprints) that can be used as bottom layer cases. Each simulation also has a collection fo "Basic stacks", each containing a runnable case and optionally a template. The editor allows to pick the template of the basic cases among the simulation's templates, but technically, a template can be any case of the workspace.

Case sharing

Blueprints can share cases. This automatically occurs when a blueprint is duplicated, but not when a blueprint is exported and reimported where each case is duplicated.

When cases are shared, they can be modified at one for every blueprint. This is useful when you want to fix an error.

Workspace cases collections

Cases

Cases are grouped in a multi-layer stack. Each layer is a single case.

The content of a case can be used as a base for another case, forming a stack (technically, even a single case is part of a stack).

The basic stack is a two-layered stack (used as the only existing stack structure)

Cases are created workspace-wide whereas case stacks are limited to the blueprint scope. This structure allows to share cases among several blueprints.

Functions

create_blank_stack(blueprint, name)

Creates a new stack with an empty runnable case and no template case.

create_blank_template(blueprint, name)

Create a new blank template for the given blueprint with the given name.

create_input_entry(kase, key, value \\ "-")

Create a new input entry for the given case, key with the given value.

create_output_entry(kase, key, variable)

Create a new output entry for the given case, key and variable.

create_setting(blueprint)

Create a new simulation setting for the given blueprint.

The default "now" time used here is "2000-01-01" (0h).

create_stack(blueprint, attrs \\ %{})

Creates a new stack using the given attributes.

delete_input_entry(input_entry)

Delete the given input entry.

delete_output_entry(output_entry)

Delete given output entry.

delete_stack(stack_id)

Deletes the stack with the given id. Layers are also deleted. Cases not.

delete_stack_template(stack)

Delete the template of the given simple stack.

delete_template(template_id)

Delete the template with the given id.

fork_runnable_case(stack, name)

Fork the runnable case of the given simple stack, i.e. create a copy to use in the given stack.

This is useful when the runnable case is shared among stacks of several blueprints, e.g. when a blueprint has been duplicated.

fork_template_case(template, name)

For the template case, i.e. create a copy to use in the template and all cases using this template.

This is useful when the template case is shared among templates of several blueprints, e.g. when a blueprint has been duplicated.

get_blueprints_sharing_runnable_case(stack)

Get the blueprint name and ids that share the runnable case of the given simple (two-layer) stack.

get_blueprints_sharing_template_case(template)

Returns the name and ids of the blueprints with which the case is shared.

get_cases_using_template(template)

Retrieve the cases (stacks) using the given template.

get_first_stack(blueprint)

Get the first stack associated to the given blueprint.

get_setting(blueprint)

Get the simulation settings.

get_stack(stack_id)

Retrieve a stack corresponding to the given id.

get_stack!(stack_id, queries)

Retrieve a stack corresponding to the given id (throw! version).

get_stack_names(blueprint)

Retrieve all the stack names for the given blueprint.

Typically used to build the list of available stacks.

get_stack_runnable_case(stack)

Get the runnable case of the given simple stack.

get_stack_template_case(stack)

Get the template case of the given simple stack.

get_stacks(blueprint)

Get all stacks for the given blueprint.

get_template(template_id)

Retrieve the template with the given id.

get_template_names(blueprint)

Get all the given blueprint's template names.

get_templates(blueprint)

Get all the given blueprint's templates.

import_all_cases(path)

Import all cases from a path (used for bulk importation).

load_stack_layers_case_entries(query)

Add the case entries (input and output) to the stacks retrieved by the given query.

load_stack_setting(query)

Add the setting attribute to the stacks retrieved by the given query.

queue_job(job)

Queue a case to be runned by the simulation runner.

set_expect_run_error(kase, expect_run_error)

Sets whether the case expects an error to occur (mostly when the complete element is not used).

set_expected(entry, expected)

Set the "expected" value of the given output entry.

set_stack_template(stack, template_case_id)

Set the case with the given id as template case for the given simple stack.

toggle_forbidden(entry, forbidden)

Set the forbidden flag of the given output entry. If forbidden is set to true, the corresponding output variable is expected to be absent of the output in the current case.

update_input_entry(input_entry, value, variable)

Update given input entry value.

update_setting(setting, options)

Update setting with the given options. Currently, the only option is the datetime env_now.

validate_input_entry(changeset, variable)

Validate given input entry for the given variable.

validate_setting(changeset)

Validate the setting.

variable_default_value(type, enum)

Get the default value for a variable of the given type. If the type is string, the enum must me passed if it exists.