VacEngine.Simulation (Vac Engine v1.0.3)

Provides the model for the blueprint simulation. Simulation is useful to assess whether a ruleset meets the business requirements.

Simulation is based on cases that describe the expected output for a given input.

The content of a case can be used as a base for another case, forming a stack (technically, even a single case is part of a stack).

Cases are created workspace-wide whereas case stacks are limited to the blueprint scope. This structure allows to share cases among several blueprints.

Functions

create_blank_output_entry(kase, key, variable)

create_blank_stack(blueprint, name)

create_blank_template(blueprint, name)

create_input_entry(kase, key, value \\ "-")

create_setting(blueprint)

create_stack(blueprint, attrs \\ %{})

delete_input_entry(input_entry)

delete_output_entry(output_entry)

delete_stack(stack_id)

delete_stack_template(stack)

delete_template(template_id)

do_import_all_cases(data)

filter_cases_by_workspace(query, workspace)

filter_stacks_by_blueprint(query, blueprint)

get_case(case_id, queries)

get_case!(case_id, queries)

get_first_stack(blueprint)

get_setting(blueprint)

get_stack(stack_id)

get_stack(stack_id, queries)

get_stack!(stack_id, queries)

get_stack_names(blueprint)

get_stack_runnable_case(stack)

get_stack_template_case(stack)

get_stacks(blueprint)

get_template(template_id)

get_template_names(blueprint)

get_templates(blueprint)

import_all_cases(path)

list_cases(queries)

list_stacks(queries)

load_stack_layers(query)

load_stack_layers_case_entries(query)

load_stack_setting(query)

queue_job(job)

set_expect_run_error(kase, expect_run_error)

set_expected(entry, expected)

set_stack_template(stack, template_case_id)

toggle_forbidden(entry, forbidden)

update_input_entry(input_entry, value)

update_setting(setting, list)

validate_input_entry(changeset, variable)

validate_setting(changeset)

variable_default_value(type, enum)