VacEngineWeb.Router.Helpers (Vac Engine v1.0.1)

Module with named helpers generated from VacEngineWeb.Router.

Functions

api_key_path(conn_or_endpoint, action)

api_key_path(conn_or_endpoint, action, params)

api_key_path(conn_or_endpoint, action, role_id, params)

api_key_url(conn_or_endpoint, action)

api_key_url(conn_or_endpoint, action, params)

api_key_url(conn_or_endpoint, action, role_id, params)

api_pub_path(conn_or_endpoint, action, portal_id)

api_pub_path(conn_or_endpoint, action, portal_id, params)

api_pub_url(conn_or_endpoint, action, portal_id)

api_pub_url(conn_or_endpoint, action, portal_id, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

login_path(conn_or_endpoint, action)

login_path(conn_or_endpoint, action, params)

login_path(conn_or_endpoint, action, token, params)

login_url(conn_or_endpoint, action)

login_url(conn_or_endpoint, action, params)

login_url(conn_or_endpoint, action, token, params)

logout_path(conn_or_endpoint, action)

logout_path(conn_or_endpoint, action, params)

logout_url(conn_or_endpoint, action)

logout_url(conn_or_endpoint, action, params)

nav_path(conn_or_endpoint, action)

nav_path(conn_or_endpoint, action, params)

nav_url(conn_or_endpoint, action)

nav_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, user_id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, user_id, params)

welcome_path(conn_or_endpoint, action)

welcome_path(conn_or_endpoint, action, params)

welcome_url(conn_or_endpoint, action)

welcome_url(conn_or_endpoint, action, params)

workspace_blueprint_path(conn_or_endpoint, action, workspace_id)

workspace_blueprint_path(conn_or_endpoint, action, workspace_id, params)

workspace_blueprint_path(conn_or_endpoint, action, workspace_id, blueprint_id, params)

workspace_blueprint_url(conn_or_endpoint, action, workspace_id)

workspace_blueprint_url(conn_or_endpoint, action, workspace_id, params)

workspace_blueprint_url(conn_or_endpoint, action, workspace_id, blueprint_id, params)

workspace_dashboard_path(conn_or_endpoint, action, workspace_id)

workspace_dashboard_path(conn_or_endpoint, action, workspace_id, params)

workspace_dashboard_url(conn_or_endpoint, action, workspace_id)

workspace_dashboard_url(conn_or_endpoint, action, workspace_id, params)

workspace_path(conn_or_endpoint, action)

workspace_path(conn_or_endpoint, action, params)

workspace_path(conn_or_endpoint, action, workspace_id, params)

workspace_portal_path(conn_or_endpoint, action, workspace_id)

workspace_portal_path(conn_or_endpoint, action, workspace_id, params)

workspace_portal_path(conn_or_endpoint, action, workspace_id, portal_id, params)

workspace_portal_url(conn_or_endpoint, action, workspace_id)

workspace_portal_url(conn_or_endpoint, action, workspace_id, params)

workspace_portal_url(conn_or_endpoint, action, workspace_id, portal_id, params)

workspace_url(conn_or_endpoint, action)

workspace_url(conn_or_endpoint, action, params)

workspace_url(conn_or_endpoint, action, workspace_id, params)